# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AskRequest": {
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "answered": {
            "title": "Answered",
            "type": "boolean"
          },
          "table": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AskTable"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "answered",
          "answer"
        ],
        "title": "AskResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryItem": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "integer"
          },
          "days_of_cover": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Cover"
          },
          "on_the_way": {
            "title": "On The Way",
            "type": "integer"
          },
          "product": {
            "$ref": "#/components/schemas/ProductRef"
          },
          "runs_low_day": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runs Low Day"
          },
          "spare_units": {
            "title": "Spare Units",
            "type": "integer"
          },
          "status": {
            "enum": [
              "RUNNING_LOW",
              "EXTRA",
              "OK"
            ],
            "title": "Status",
            "type": "string"
          },
          "store": {
            "$ref": "#/components/schemas/Ref"
          }
        },
        "required": [
          "store",
          "product",
          "status",
          "available",
          "on_the_way",
          "days_of_cover",
          "runs_low_day",
          "spare_units"
        ],
        "title": "InventoryItem",
        "type": "object"
      },
      "MarkdownDecision": {
        "properties": {
          "decision": {
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "title": "Decision",
            "type": "string"
          },
          "product_id": {
            "pattern": "^P\\d{2}$",
            "title": "Product Id",
            "type": "string"
          },
          "store_id": {
            "pattern": "^S\\d{2}$",
            "title": "Store Id",
            "type": "string"
          }
        },
        "required": [
          "store_id",
          "product_id",
          "decision"
        ],
        "title": "MarkdownDecision",
        "type": "object"
      },
...
      "ProductRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "name_plural": {
            "title": "Name Plural",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "name_plural"
        ],
        "title": "ProductRef",
        "type": "object"
      },
...
      "StormDeskRequest": {
        "properties": {
          "goal": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Goal",
            "type": "string"
          }
        },
        "required": [
          "goal"
        ],
        "title": "StormDeskRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "StormSense",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ask": {
      "post": {
        "operationId": "ask_api_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "ask"
        ]
      }
    },
...
    "/api/inventory": {
      "get": {
        "operationId": "inventory_api_inventory_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "RUNNING_LOW",
                    "EXTRA"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  },
                  "title": "Response Inventory Api Inventory Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inventory",
        "tags": [
          "stores"
        ]
      }
    },
    "/api/markdowns": {
      "get": {
        "description": "Surplus stock that would not sell in time at full price, with a discount that adds cash. Suggestions only: nothing changes.",
        "operationId": "markdowns_api_markdowns_get",
        "parameters": [
          {
            "in": "query",
            "name": "weather",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "demo"
              ],
              "title": "Weather",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarkdownSuggestion"
                  },
                  "title": "Response Markdowns Api Markdowns Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Markdowns",
        "tags": [
          "stores"
        ]
      }
    },
    "/api/markdowns/decision": {
      "post": {
        "description": "A planner approves or rejects a markdown suggestion. Records the decision; nothing changes in the stores.",
        "operationId": "decide_markdown_api_markdowns_decision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkdownDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkdownSuggestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Markdown",
        "tags": [
          "stores"
        ]
      }
    },
...
    "/api/overview": {
      "get": {
        "operationId": "overview_api_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "weather",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "demo"
              ],
              "title": "Weather",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "today"
        ]
      }
    },
    "/api/stores": {
      "get": {
        "operationId": "stores_api_stores_get",
        "parameters": [
          {
            "in": "query",
            "name": "weather",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "demo"
              ],
              "title": "Weather",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StoreSummary"
                  },
                  "title": "Response Stores Api Stores Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stores",
        "tags": [
          "stores"
        ]
      }
    },
    "/api/stores/{store_id}/forecast": {
      "get": {
        "operationId": "store_forecast_api_stores__store_id__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "store_id",
            "required": true,
            "schema": {
              "pattern": "^S\\d{2}$",
              "title": "Store Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "weather",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "demo"
              ],
              "title": "Weather",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreForecast"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Forecast",
        "tags": [
          "stores"
        ]
      }
    },
    "/api/storm-desk": {
      "post": {
        "description": "Plans from the live data. Read-only: it can suggest moves but never approves or changes anything.",
        "operationId": "storm_desk_api_storm_desk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StormDeskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormDeskPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storm Desk",
        "tags": [
          "storm desk"
        ]
      }
    },
    "/api/transfers": {
      "get": {
        "operationId": "transfers_api_transfers_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "APPROVED",
                    "REJECTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "urgency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "URGENT",
                    "NORMAL"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Urgency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Transfer"
                  },
                  "title": "Response Transfers Api Transfers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfers",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/approve": {
      "post": {
        "operationId": "approve_api_transfers_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/export": {
      "get": {
        "description": "The plan as a spreadsheet (CSV). Read-only. Open it in a spreadsheet, or print it to PDF from the browser.",
        "operationId": "export_transfers_api_transfers_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "APPROVED",
                    "REJECTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Transfers",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/reject": {
      "post": {
        "operationId": "reject_api_transfers_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/{transfer_id}": {
      "get": {
        "operationId": "transfer_api_transfers__transfer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transfer_id",
            "required": true,
            "schema": {
              "pattern": "^TR-[A-Z0-9]{10}$",
              "title": "Transfer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/what-if": {
      "post": {
        "description": "Simulates a storm and what it would cost. Read-only: it changes no data and approves nothing.",
        "operationId": "what_if_api_what_if_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What If",
        "tags": [
          "what-if"
        ]
      }
    }
  }
}